float version is stable

+++ NeuralNet/ANN.h
@@ -20,47 +20,47 @@ typedef struct _MSI *MSLink;
 
 /*model set info -this struct is needed to compare the error on the validation dataset between two epochs*/
 typedef struct _MSI{
-double crtVal;
-double bestValue;
-double prevCrtVal;
+float crtVal;
+float bestValue;
+float prevCrtVal;
 }MSI;
 
 /**this struct stores the search directions,residues and dw for each iteration of CG**/
 typedef struct _ConjugateGradientInfo{
-   double *delweightsUpdate;
-   double *delbiasUpdate;
-   double *residueUpdateWeights;
-   double * residueUpdateBias;
-   double *searchDirectionUpdateWeights;
-   double *searchDirectionUpdateBias;
+   float *delweightsUpdate;
+   float *delbiasUpdate;
+   float *residueUpdateWeights;
+   float * residueUpdateBias;
+   float *searchDirectionUpdateWeights;
+   float *searchDirectionUpdateBias;
  }ConjuageGradientInfo;
 
 
 
 /** this struct stores the directional error derivatives with respect to weights and biases**/
 typedef struct _GaussNewtonProdInfo{
-	double *vweights;
-   	double *vbiases;
-   	double *Ractivations;
+	float *vweights;
+   	float *vbiases;
+   	float *Ractivations;
 }GaussNewtonProductInfo ;
 
 typedef struct _TrainInfo{
-	double *dwFeatMat; /* dE/dw matrix*/
-	double *dbFeaMat; /* dE/db  vector */
-	double *updatedWeightMat; /* stores the velocity in the weight space or accumulates  gradeints of weights*/
-	double *updatedBiasMat;/* stores the velocity in the bias space  or accumulates gradients of biases*/
-	double *bestWeightParamsHF;
-	double *bestBiasParamsHF;
+	float *dwFeatMat; /* dE/dw matrix*/
+	float *dbFeaMat; /* dE/db  vector */
+	float *updatedWeightMat; /* stores the velocity in the weight space or accumulates  gradeints of weights*/
+	float *updatedBiasMat;/* stores the velocity in the bias space  or accumulates gradients of biases*/
+	float *bestWeightParamsHF;
+	float *bestBiasParamsHF;
 }TrainInfo;
 
 typedef struct _ErrorElem{
-	double *dxFeatMat;
-	double *dyFeatMat;
+	float *dxFeatMat;
+	float *dyFeatMat;
 }ErrElem;
 
 typedef struct _FeatElem{
-	double *xfeatMat;/* is a BatchSample size by feaDim matix */
-	double *yfeatMat; /* is BatchSample Size by node matrix*/
+	float *xfeatMat;/* is a BatchSample size by feaDim matix */
+	float *yfeatMat; /* is BatchSample Size by node matrix*/
 }FeaElem;
 
 /*structure for individual layers*/
@@ -71,15 +71,15 @@ typedef struct _LayerElem{
 	ActFunKind actfuncKind; /*each layer is allowed to have its own non-linear activation function */
 	LELink src; /* pointer to the input layer */
 	int srcDim; /* the number of units in the input layer */
-	double *weights;/* the weight matrix of the layer should number of nodes by input dim*/
-	double *bias; /* the bias vector */
+	float *weights;/* the weight matrix of the layer should number of nodes by input dim*/
+	float *bias; /* the bias vector */
 	FELink feaElem; /* stores the  input activations coming into the layer as well the output activations going out from the layer */
 	ERLink errElem;
 	TRLink traininfo;/*struct that stores the error derivatives with respect to weights and biases */
 	GNProdInfo gnInfo;
 	CGInfo  cgInfo;
-	double *bestweights;
-	double *bestBias;
+	float *bestweights;
+	float *bestBias;
 }LayerElem;
 
 /*structure for ANN*/
@@ -88,15 +88,15 @@ typedef struct _ANNdef{
 	LELink *layerList;/* list of layers*/
 	OutFuncKind target; /* the activation function of the final output layer */
 	ObjFuncKind errorfunc; /* the error function of the ANN */
-	double *labelMat ; /* the target labels : BatchSample by targetDim matrix*/
+	float *labelMat ; /* the target labels : BatchSample by targetDim matrix*/
 }ANNDef;
 
 //-------------------------------------------------------------------------------------------------------------------
 /**This section of the code deals with parsing Command Line arguments**/
 //-------------------------------------------------------------------------------------------------------------------
 void cleanString(char *Name);
-void loadLabels(double *labelMat, int *labels,char*filepath,char *datatype);
-void loadMatrix(double *matrix,char *filepath, char *datatype);
+void loadLabels(float *labelMat, int *labels,char*filepath,char *datatype);
+void loadMatrix(float *matrix,char *filepath, char *datatype);
 void parseCfg(char * filepath);
 void parseCMDargs(int argc, char *argv[]);
 
@@ -108,7 +108,7 @@ void setBatchSize(int sampleSize);
 void setBatchSizetoHFminiBatch();
 void loadMiniBatchintoANN();
 /**load entire batch into the neural net**/
-void loadDataintoANN(double *samples, double *labels);
+void loadDataintoANN(float *samples, float *labels);
 
 //-------------------------------------------------------------------------------------------------------------------
 /**this section of the src code deals with initialisation of ANN **/
@@ -119,14 +119,14 @@ void setUpForHF(ADLink anndef);
 void reinitLayerFeaMatrices(ADLink anndef);
 void reinitLayerErrFeaMatrices(ADLink anndef);
 void initialiseErrElems(ADLink anndef);
-void initialiseWithZero(double * matrix, int dim);
-double drand();
-double genrandWeight(double limit);
-void initialiseBias(double *biasVec,int dim, int srcDim,ActFunKind actfunc);
+void initialiseWithZero(float * matrix, int dim);
+float drand();
+float genrandWeight(float limit);
+void initialiseBias(float *biasVec,int dim, int srcDim,ActFunKind actfunc);
 /*the srcDim determines the fan-in to the hidden and output units. The weights ae initialised 
 to be inversely proportional to the sqrt(fanin)
 */ 
-void initialiseWeights(double *weightMat,int length,int srcDim,ActFunKind actfunc);
+void initialiseWeights(float *weightMat,int length,int srcDim,ActFunKind actfunc);
 void initialiseLayer(LELink layer,int i, LELink srcLayer);
 void  initialiseDNN();
 void initialise();
@@ -135,13 +135,13 @@ void initialise();
 /* this section of the code presents auxilary functions that are required*/
 //-------------------------------------------------------------------------------------------------------------------
 /**this funciton copies one matrix/array into another*/
-void copyMatrixOrVec(double *src, double *dest,int dim);
+void copyMatrixOrVec(float *src, float *dest,int dim);
 /* this function allows the addition of  two matrices or two vectors*/
-void addMatrixOrVec(double *weightMat, double* dwFeatMat, int dim,double lambda);
-void scaleMatrixOrVec(double* weightMat, double learningrate,int dim);
-void subtractMatrix(double *dyfeat, double* labels, int dim, double lambda);
-double computeTanh(double x);
-double computeSigmoid(double x);
+void addMatrixOrVec(float *weightMat, float* dwFeatMat, int dim,float lambda);
+void scaleMatrixOrVec(float* weightMat, float learningrate,int dim);
+void subtractMatrix(float *dyfeat, float* labels, int dim, float lambda);
+float computeTanh(float x);
+float computeSigmoid(float x);
 
 //-------------------------------------------------------------------------------------------------------------------
 /*this section of the code implements the  forward propagation of a deep neural net **/
@@ -149,13 +149,13 @@ double computeSigmoid(double x);
 void computeNonLinearActOfLayer(LELink layer);
 /* Yfeat is batchSamples by nodeNum matrix(stored as row major)  = X^T(row-major)-batch samples by feaMat * W^T(column major) -feaMat By nodeNum */
 void computeLinearActivation(LELink layer);
-void loadDataintoANN(double *samples, double *labels);
+void loadDataintoANN(float *samples, float *labels);
 void fwdPassOfDNN(ADLink anndef);
 void computeActivationOfOutputLayer(ADLink anndef);
 //------------------------------------------------------------------------------------------------------
 /*This section of the code implements the back-propation algorithm  to compute the error derivatives*/
 //-------------------------------------------------------------------------------------------------------
-void sumColsOfMatrix(double *dyFeatMat,double *dbFeatMat,int dim,int batchsamples);
+void sumColsOfMatrix(float *dyFeatMat,float *dbFeatMat,int dim,int batchsamples);
 void computeActivationDrv (LELink layer);
 /**compute del^2L J where del^2L is the hessian of the cross-entropy softmax with respect to output acivations **/ 
 void computeLossHessSoftMax(LELink layer);
@@ -171,15 +171,15 @@ void backPropBatch(ADLink anndef,Boolean doHessVecProd);
 void fillCache(LELink layer,int dim,Boolean weights);
 void cacheParameters(ADLink anndef);
 Boolean initialiseParameterCaches(ADLink anndef);
-void perfBinClassf(double *yfeatMat, double *predictions,int dataSize);
-double computeLogLikelihood(double* output, int batchsamples, int dim , int* labels);
+void perfBinClassf(float *yfeatMat, float *predictions,int dataSize);
+float computeLogLikelihood(float* output, int batchsamples, int dim , int* labels);
 /*The function finds the most active node in the output layer for each sample*/
-void findMaxElement(double *matrix, int row, int col, double *vec);
+void findMaxElement(float *matrix, int row, int col, float *vec);
 /** the function calculates the percentage of the data samples correctly labelled by the DNN*/
-double updatateAcc(int *labels, LELink layer,int dataSize);
-void updateNeuralNetParams(ADLink anndef, double lrnrate, double momentum, double weightdecay);
-void updateLearningRate(int currentEpochIdx, double *lrnRate);
-Boolean terminateSchedNotTrue(int currentEpochIdx,double lrnrate);
+float updatateAcc(int *labels, LELink layer,int dataSize);
+void updateNeuralNetParams(ADLink anndef, float lrnrate, float momentum, float weightdecay);
+void updateLearningRate(int currentEpochIdx, float *lrnRate);
+Boolean terminateSchedNotTrue(int currentEpochIdx,float lrnrate);
 void TrainDNNGD();
 
 //------------------------------------------------------------------------------------------------------
@@ -187,7 +187,7 @@ void TrainDNNGD();
 //------------------------------------------------------------------------------------------------------
 void setHook(Ptr m, Ptr ptr,int incr);
 Ptr getHook(Ptr m,int incr);
-void accumulateLayerGradient(LELink layer,double weight);
+void accumulateLayerGradient(LELink layer,float weight);
 void accumulateGradientsofANN(ADLink anndef);
 
 //------------------------------------------------------------------------------------------------------
@@ -195,7 +195,7 @@ void accumulateGradientsofANN(ADLink anndef);
 //------------------------------------------------------------------------------------------------------
 void normOfWeights(ADLink anndef);
 void normaliseSearchDirections(ADLink anndef);
-void normaliseResidueDirections(ADLink anndef, double* magnitudeOfGradient);
+void normaliseResidueDirections(ADLink anndef, float* magnitudeOfGradient);
 void computeNormOfGradient(ADLink anndef);
 void computeNormOfAccuGradient(ADLink anndef);
 void normOfVweights(ADLink anndef);
@@ -219,12 +219,12 @@ void backtrackNeuralNetParamsCG(ADLink anndef);
 
 //-------------------------------------------------------
 /**This section of the code implements the small sub routinesof the  conjugate Gradient algorithm **/
-void updateParameterDirection(ADLink anndef,double beta);
+void updateParameterDirection(ADLink anndef,float beta);
 void updateResidue(ADLink anndef);
-double  computeQuadfun( ADLink anndef);
-void updatedelParameters(double alpha);
-void computeSearchDirMatrixProduct( ADLink anndef,double * searchVecMatrixVecProductResult);
-void computeResidueDotProduct(ADLink anndef, double * residueDotProductResult);
+float  computeQuadfun( ADLink anndef);
+void updatedelParameters(float alpha);
+void computeSearchDirMatrixProduct( ADLink anndef,float * searchVecMatrixVecProductResult);
+void computeResidueDotProduct(ADLink anndef, float * residueDotProductResult);
 void addTikhonovDamping(ADLink anndef);
 //------------------------------------------------------
 /* the following routines compute the directional derivative using forward differentiation*/
@@ -233,7 +233,7 @@ void computeRactivations(LELink layer);
 void computeVweightsProjection(LELink layer);
 void computeDirectionalErrDrvOfLayer(LELink layer, int layerid);
 void computeDirectionalErrDerivativeofANN(ADLink anndef);
-void setParameterDirections(double * weights, double* bias, LELink layer);
+void setParameterDirections(float * weights, float* bias, LELink layer);
 void setSearchDirectionCG(ADLink anndef, Boolean Parameter);
 //-----------------------------------------------------
 void initialiseResidueaAndSearchDirection(ADLink anndef);
@@ -247,11 +247,11 @@ void TrainDNNHF();
 void freeMemoryfromANN();
 void printLayerMatrices(ADLink anndef);
 void printLayerDWMatrices(ADLink anndef);
-void printVector(double * vector , int dim);
+void printVector(float * vector , int dim);
 void UnitTest_computeGradientDotProd();
 void printWeights(ADLink anndef, int i);
 void printYfeat(ADLink anndef, int id);
-void printMatrix(double *matrix,int row,int col);
+void printMatrix(float *matrix,int row,int col);
 void printDBWeights(ADLink anndef, int i);
 
 /*This function is used to check the correctness of various routines */
